refactor: RegisterForm 추상화
