+++ src/view/group/problem-list/ProblemSidebar/index.tsx
@@ -1,13 +1,36 @@
 "use client";
+import { usePostProblemMutation } from "@/app/group/[groupId]/problem-list/query";
 import Button from "@/common/component/Button";
 import Modal from "@/common/component/Modal";
 import { useBooleanState } from "@/common/hook/useBooleanState";
+import useGetGroupId from "@/shared/hook/useGetGroupId";
 import { problemSidebarWrapper } from "@/view/group/problem-list/ProblemSidebar/index.css";
 import RegisterForm from "@/view/group/problem-list/RegisterForm";
+import type { Dispatch, SetStateAction } from "react";
 
 const ProblemSidebar = () => {
   const { open, isOpen, close } = useBooleanState();
+  const { mutate: postProblemMutate } = usePostProblemMutation();
+  const groupId = useGetGroupId();
 
+  const handleSubmit = (
+    link: string,
+    startDate: Date,
+    endDate: Date,
+    setIsSuccess: Dispatch<SetStateAction<boolean>>,
+  ) => {
+    postProblemMutate(
+      { groupId: +groupId, link, startDate, endDate },
+      {
+        onSuccess: () => {
+          setIsSuccess(true);
+          setTimeout(() => {
+            close();
+          }, 1700);
+        },
+      },
+    );
+  };
   return (
     <>
       <div className={problemSidebarWrapper}>
@@ -16,7 +39,7 @@ const ProblemSidebar = () => {
         </Button>
       </div>
       <Modal isOpen={isOpen} onClose={close}>
-        <RegisterForm onRegister={close} />
+        <RegisterForm onSubmit={handleSubmit} />
       </Modal>
     </>
   );
